Cleanup Account and Authorization
 Rename ULC to use Account instead of User for clarity
 Move User repository to Authorization and out of ULC*
 Move Firebase Auth code to ULC
 Add some template code for other issues
