+++ web_purduehcr/lib/Account_Login_Creation/CreateAccountPage.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+
+class CreateAccountPage extends StatelessWidget {
+  const CreateAccountPage({Key key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        backgroundColor: Color.fromARGB(255, 220, 220, 220),
+        body: Center(
+          child: Container(
+            width: MediaQuery.of(context).size.width * 0.22,
+            height: MediaQuery.of(context).size.height * 0.4,
+            child: Text("Unimplemented"),
+          ),
+        )
+    );
+  }
+}
