--- web_purduehcr/lib/BLoCs/authentication/authentication_event.dart
@@ -1,32 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
-import 'package:purduehcr_web/Models/User.dart';
-
-abstract class AuthenticationEvent extends Equatable {
-  const AuthenticationEvent();
-
-  @override
-  List<Object> get props => [];
-}
-
-class AppStarted extends AuthenticationEvent {
-
-  const AppStarted();
-
-  @override
-  List<Object> get props => [];
-}
-
-class LoggedIn extends AuthenticationEvent {
-  final User user;
-
-  const LoggedIn({@required this.user});
-
-  @override
-  List<Object> get props => [user];
-
-  @override
-  String toString() => 'LoggedIn { user: $user }';
-}
-
-class LoggedOut extends AuthenticationEvent {}
