+++ web_purduehcr/lib/OverviewPage/ResidentOverviewWidget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:purduehcr_web/BLoCs/authentication/authentication.dart';
+import 'package:purduehcr_web/authentication/authentication.dart';
 import 'package:purduehcr_web/OverviewPage/overview_bloc/overview.dart';
 import 'package:purduehcr_web/Models/User.dart';
 import 'package:purduehcr_web/OverviewPage/overview_cards/ProfileCard.dart';
