+++ web_purduehcr/lib/PurdueHCR_main.dart
@@ -4,9 +4,8 @@ import 'package:purduehcr_web/ConfigWrapper.dart';
 import 'package:purduehcr_web/RouteGenerator.dart';
 import 'package:bloc/bloc.dart';
 
-import 'package:purduehcr_web/BLoCs/authentication/authentication.dart';
+import 'package:purduehcr_web/authentication/authentication.dart';
 
-import 'User_Login_Creation/user_login_creation_bloc/ulc_repository.dart';
 
 
 class SimpleBlocDelegate extends BlocDelegate {
