+++ web_purduehcr/lib/RouteGenerator.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:purduehcr_web/ConfigWrapper.dart';
 import 'package:purduehcr_web/OverviewPage/OverviewPage.dart';
-import 'package:purduehcr_web/User_Login_Creation/LogInPage.dart';
+import 'package:purduehcr_web/Account_Login_Creation/LogInPage.dart';
 import 'package:purduehcr_web/TokenTestPage/TokenTestPage.dart';
 
-import 'package:purduehcr_web/BLoCs/authentication/authentication.dart';
+import 'package:purduehcr_web/authentication/authentication.dart';
 
 class RouteGenerator {
   static Route<dynamic> generateRoute(RouteSettings settings) {
