--- web_purduehcr/lib/User_Login_Creation/user_login_creation_bloc/ulc_event.dart
@@ -1,30 +1,0 @@
-
-import 'package:equatable/equatable.dart';
-import 'package:flutter/material.dart';
-
-abstract class ULCEvent extends Equatable {
-  const ULCEvent();
-}
-
-class ULCInitialize extends ULCEvent {
-  const ULCInitialize();
-  @override
-  List<Object> get props => [];
-}
-
-class Login extends ULCEvent {
-
-  final String email;
-  final String password;
-  const Login({
-    @required this.email,
-    @required this.password,
-  });
-  @override
-  List<Object> get props => [email, password];
-
-  @override
-  String toString() =>
-      'ULCButtonPressed { username: $email, password: $password }';
-
-}
