+++ web_purduehcr/lib/authentication/authentication_repository.dart
@@ -1,0 +1,25 @@
+import 'package:purduehcr_web/Config.dart';
+import 'package:purduehcr_web/Models/User.dart';
+import 'package:purduehcr_web/Utilities/CloudFunctionUtility.dart';
+import 'package:purduehcr_web/Utilities/FirebaseUtility.dart';
+
+class AuthenticationRepository {
+  final Config config;
+
+  AuthenticationRepository(this.config);
+
+
+  Future<String> createUser(String first, String last, String code) {
+    // TODO: implement createUser
+    throw UnimplementedError();
+  }
+
+  Future<void> logout(){
+    return FirebaseUtility.logout();
+  }
+
+  Future<User> getUser() async {
+    Map<String, dynamic> userMap = await callCloudFunction(config, Method.GET, "user/get");
+    return User.fromJson(userMap);
+  }
+}
